+++ CHANGELOG.rst
@@ -7,8 +7,10 @@ Changelog
 
 Bugfixes
 --------
-* Fix ``MostRecentProvider`` behavior when lock cannot be acquired.
+* Fix :class:`MostRecentProvider` behavior when lock cannot be acquired.
   `#72 <https://github.com/awslabs/aws-dynamodb-encryption-python/issues/72>`_
+* Fix :class:`MostRecentProvider` lock acquisition for Python 2.7.
+  `#73 <https://github.com/awslabs/aws-dynamodb-encryption-python/issues/73>`_
 
 1.0.3 -- 2018-05-03
 ===================